Move ConvoBot.java to cli folder and add GUI icons
