--- src/main/java/convobot/gui/Convo.java
@@ -1,44 +1,0 @@
-package gui;
-
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
-
-/**
- * The main class for the ConvoBot GUI application.
- */
-public class Convo {
-
-    // private final ConvoBot convoBot = new ConvoBot("./data/tasks.txt");
-
-    private final Image user = new Image(getClass().getResourceAsStream("/images/DaUser.png"));
-    private final Image duke = new Image(getClass().getResourceAsStream("/images/DaDuke.png"));
-
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    private Scene scene;
-
-    public Convo() {
-        // ...
-    }
-
-    /**
-     * You should have your own function to generate a response to user input.
-     * Replace this stub with your completed method.
-     */
-    String getResponse(String input) {
-        return "Duke heard: " + input;
-    }
-
-}
