+++ src/main/java/convobot/gui/Main.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
+import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
@@ -13,16 +14,16 @@ import javafx.stage.Stage;
  */
 public class Main extends Application {
 
-    private Convo convo = new Convo();
-
     @Override
     public void start(Stage stage) {
+        stage.setTitle("ConvoBot");
+        stage.getIcons().add(new Image(getClass().getResourceAsStream("/images/DaConvo.png")));
+
         try {
             FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setConvo(convo);
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
